Move internal helpers into separate modules
By not exporting these, and automatically excluding them from the
generated documentation, we can be sure that they are not used by
developers using a recent version of Node (12+) or Webpack,
because they are not listed in the exports map.

Note that, since exports maps are not yet supported by most of the
ecosystem (e.g. Webpack <5, Node 10), which hence still allow
developers to import *anything* we export,  I preserved the
`internal_` prefix, to make it clear that those should not be
imported even if you can, because we can change their API's at any
time.
