+++ src/interfaces.ts
@@ -219,11 +219,6 @@ export type WithFallbackAcl<
   };
 };
 
-/** @internal */
-export function internal_toIriString(iri: Iri | IriString): IriString {
-  return typeof iri === "string" ? iri : iri.value;
-}
-
 /**
  * Verify whether a given SolidDataset includes metadata about where it was sent to.
  *
